adding testing for legacy configuration
