+++ src/NHiLo.Tests.Integration/NHiLo.Tests.Integration.csproj
@@ -32,4 +32,10 @@
     <ProjectReference Include="..\NHiLo\NHiLo.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <None Update="testhost.dll.config">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
 </Project>
