+++ src/NHiLo.Tests.Integration/testhost.dll.config
@@ -1,0 +1,18 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<!-- for the configuration legacy tests -->
+<configuration>
+	<configSections>
+		<section name="nhilo" type="NHiLo.Common.Legacy.KeyGeneratorConfig, NHiLo"/>
+	</configSections>
+	<connectionStrings>
+		<add name="DB1" connectionString="test" providerName="myRepository"/>
+	</connectionStrings>
+	<nhilo>
+		<hiloKeyGenerator connectionStringId="DB1" createHiLoStructureIfNotExists="true" defaultMaxLo="10" storageType="Sequence" objectPrefix="SQXX_">
+			<entities>
+				<add name="xpto1" maxLo="100" />
+				<add name="xpto2" maxLo="66" />
+			</entities>
+		</hiloKeyGenerator>
+	</nhilo>
+</configuration>
